fix(eval): load tokenizer from TEACHER_MODEL; explicit dtype=torch.long for ctx
save_checkpoint() calls model.save_pretrained() which saves weights and config.json
but NOT tokenizer files (tokenizer.json, tokenizer_config.json, etc.).
AutoTokenizer.from_pretrained(<checkpoint_dir>) with no tokenizer files can silently
fall back to a broken tokenizer whose .encode() returns floats, causing:

  RuntimeError: Expected tensor for argument #1 'indices' to have scalar types:
  Long, Int; but got torch.cuda.FloatTensor instead

Fix 1 (eval.py): load tokenizer from TEACHER_MODEL ("Qwen/Qwen3-8B") — draft and
teacher share a vocabulary (required for speculative decoding), so this is always
the correct source regardless of the checkpoint path.

Fix 2 (eval.py, train.py): add dtype=torch.long when constructing ctx/prompt_ids
from tokenizer.encode() output — defensive guard that prevents this class of error
even if the tokenizer returns an unexpected numeric type.

+++ train.py
@@ -99,7 +99,7 @@ WANDB_PROJECT   = "distillspec-pipeline"
 
 def _prompt_to_ids(prompt: str, tokenizer, device) -> torch.Tensor:
     """Tokenise a string prompt to [1, plen] LongTensor on the model's device."""
-    return torch.tensor(tokenizer.encode(prompt), device=device).unsqueeze(0)
+    return torch.tensor(tokenizer.encode(prompt), device=device, dtype=torch.long).unsqueeze(0)
 
 
 def _prepare_prompt_cache(model, prompt_ids):
